Widen GameTypeId from byte to uint
Drops the [0, 255] cap on GameType<i>Id; replaces it with a type-level
non-negativity guarantee via uint. Parser rejects negative ids with a Warn +
skip; downstream call sites no longer need a runtime range check. Persist
blob bumps to v2 (uint32 game-type ids) with v1 byte reads preserved for
existing rating data.

+++ src/ClashEngine.Core/GameType/GameTypeDef.cs
@@ -7,13 +7,13 @@ namespace ClashEngine.Core.GameType;
 /// <summary>
 /// Pure-data view of one configured game type. Produced by the host's [ClashEngine] parser and
 /// stored in <see cref="GameTypeRegistry"/>; queues resolve their <c>GameType</c> reference
-/// through the registry by name. Two contributions sharing a name (and the same byte
-/// <see cref="Id"/>) are treated as the same logical game type for collision purposes; mismatched
-/// IDs are rejected.
+/// through the registry by name. Two contributions sharing a name (and the same
+/// <see cref="Id"/>) are treated as the same logical game type for collision purposes;
+/// mismatched IDs are rejected.
 /// </summary>
 public sealed record GameTypeDef(
     string Name,
-    byte Id,
+    uint Id,
     int TeamCount,
     int PlayersPerTeam,
     int KillTarget,
